filter out incomplete lines
in the rare cases when the logging program has not yet finished writing
its line, filehandle.next() would return an incomplete line, which in
term would be yielded by pygtail, and on the next attempt, the rest of
the line would be produced (starting at the offset).

this patch supresses incomplete lines and seeks back to the start of the
line again.