update offset file from destructor
with the follow option (but without the paranoid option), the offset
file would never be written as the end of the file is never reached.
this patch uses the Pygtail __del__ destructor to update it (eg on a
KeyboardInterrupt or system exit for other reasons).

due to the various problems associated with garbage collection and
implementation defined behavior of __del__, this patch might not be a
good idea and paranoid should be turned on by default for following
instead.